mindmapfiles_updated/requirements.txt
google-generativeai
supabase
flask-cors